:truck: 경사면에 붙는 로직 SlopeStickComponent 로 이동
