+++ Runelight-Smash-2021/Assets/Scripts/Component/SlopeStickComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e153cf66ff3d1347ba1503e05e9ee71
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 280
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
